Thu, May 11, 2023, 11:26 PM -07:00

+++ week-1/pset/credit/credit.c
@@ -260,7 +260,7 @@ int first_digit(string cc_num)
 
     long number;
     number = 0;
-    number = atoi(cc_num);
+    number = atof(cc_num);
 
     if (len == 16)
     {
